[ADD] EndProtocol transformado em componente
